Added set_pos and get_pressed to psycho mouse
Now including the runtime error when set_pos is called in a pyglet environment

Slight point of comment: now only check first window in existence, which is fine for now, but could be reconsidered if support for multiple monitors is ever added. But, now that I think of it: why would one create one pygame psychopy window and one pyglet?