When autodetect on Windows start trying from COM1
When autodetect on Windows start trying from COM1 instead of COM0 as COM0 never exists on Windows.
Note I think this will only work if the first COM port that exist actually is the SRbox. In many cases a COM1 hardware will exist, but the SRbox will be on a virtual COM port on COM 3 or 4 for example.