Fixed 'NoneType' has no len() error
Because self.timeout is now correctly passed to self._keyboard (because of Klemtonius fix: Fixed timeout function for text_input item). The while true loop tried to process the keyboard response after the timeout. Even when Accept on return press (only) was selected. No this goes well because the while loop doesn't try to get the length of a 'None' response anymore.