Added nodeappdmg.json file for DMG creation
DMG (Mac disk images) can from now on be made with node-appdmg (a
node.js package, installable with ‘npm install node-appdmg’). The
config file for this program is setup-mac-nodeappdmg.json, which sadly
cannot contain comments to elaborate on its purpose.
The einstein@2x.png is also used by node-appdmg, but it is not clear in
what sense yet.